+++ Plugins/iOS/ApphudUnityAPIBridge.m
@@ -1,94 +1,8 @@
 @import ApphudSDK;
 
 #import <Foundation/Foundation.h>
-#include "ApphudUnityPluginCallback.h"
-#include "UnityFramework/UnityFramework-Swift.h"
-
-static NSString * cstringToString(const char *str) {
-    return str ? [NSString stringWithUTF8String:str] : nil;
-}
-
-static const char * cstringFromString(NSString *str) {
-    return str ? [str cStringUsingEncoding:NSUTF8StringEncoding] : nil;
-}
-
-static char * makeStringCopy(const char *string) {
-    if (string == NULL) {
-        return NULL;
-    }
-
-    char *res = (char *)malloc(strlen(string) + 1);
-    strcpy(res, string);
-    return res;
-}
-
-static CallbackBoolDelegate _callbackBoolDelegate = NULL;
-static CallbackDelegate _callbackDelegate = NULL;
-static Callback2Delegate _callback2Delegate = NULL;
-static Callback3Delegate _callback3Delegate = NULL;
-
-void SendCallbackBoolToUnity(UnityAction action, bool data) {
-    if (action == NULL) {
-        return;
-    }
-
-    dispatch_async(dispatch_get_main_queue(), ^{
-        if (_callbackBoolDelegate != NULL) {
-            _callbackBoolDelegate(action, data);
-        }
-    });
-}
-
-void SendCallbackToUnity(UnityAction action, NSString *data) {
-    if (action == NULL) {
-        return;
-    }
-
-    dispatch_async(dispatch_get_main_queue(), ^{
-        if (_callbackDelegate != NULL) {
-            _callbackDelegate(action, cstringFromString(data));
-        }
-    });
-}
-
-void SendCallback2ToUnity(UnityAction action, NSString *data, NSString *data2) {
-    if (action == NULL) {
-        return;
-    }
-
-    dispatch_async(dispatch_get_main_queue(), ^{
-        if (_callback2Delegate != NULL) {
-            _callback2Delegate(action, cstringFromString(data), cstringFromString(data2));
-        }
-    });
-}
-
-void SendCallback3ToUnity(UnityAction action, NSString *data, NSString *data2, NSString *data3) {
-    if (action == NULL) {
-        return;
-    }
-
-    dispatch_async(dispatch_get_main_queue(), ^{
-        if (_callback3Delegate != NULL) {
-            _callback3Delegate(action, cstringFromString(data), cstringFromString(data2), cstringFromString(data3));
-        }
-    });
-}
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-void ApphudUnity_registerCallbackHandler(CallbackBoolDelegate callbackBoolDelegate,CallbackDelegate callbackDelegate, Callback2Delegate callback2Delegate, Callback3Delegate callback3Delegate) {
-    _callbackBoolDelegate = callbackBoolDelegate;
-    _callbackDelegate = callbackDelegate;
-    _callback2Delegate = callback2Delegate;
-    _callback3Delegate = callback3Delegate;
-}
-
-#ifdef __cplusplus
-}
-#endif
+#import "UnityFramework/UnityFramework-Swift.h"
+#import "ApphudUnityUtilities.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -119,7 +33,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_startWithApiKeyAndUserId(const char *apiKey,const char *userId, BOOL observerMode, UnityAction callback) {
-    [ApphudUnityWrapper startWithApiKey:cstringToString(apiKey) userID:cstringToString(userId) observerMode:observerMode callback:^(NSString * _Nonnull result) {
+    [ApphudUnityAPIWrapper startWithApiKey:cstringToString(apiKey) userID:cstringToString(userId) observerMode:observerMode callback:^(NSString * _Nonnull result) {
         SendCallbackToUnity(callback, result);
     }];
 }
@@ -133,7 +47,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_startWithApiKey(const char *apiKey, BOOL observerMode, UnityAction callback) {
-    [ApphudUnityWrapper startWithApiKey:cstringToString(apiKey) observerMode:observerMode callback:^(NSString * _Nonnull result) {
+    [ApphudUnityAPIWrapper startWithApiKey:cstringToString(apiKey) observerMode:observerMode callback:^(NSString * _Nonnull result) {
         SendCallbackToUnity(callback, result);
     }];
 }
@@ -172,7 +86,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_fetchPlacementsWithCallback(int maxAttempts, UnityAction callback) {
-    [ApphudUnityWrapper fetchPlacementsWithMaxAttempts:maxAttempts callback:^(NSString * _Nonnull placements, NSString * _Nullable error) {
+    [ApphudUnityAPIWrapper fetchPlacementsWithMaxAttempts:maxAttempts callback:^(NSString * _Nonnull placements, NSString * _Nullable error) {
         SendCallback2ToUnity(callback, placements, error);
     }];
 }
@@ -186,7 +100,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_paywallsDidLoadCallback(int maxAttempts, UnityAction callback) {
-    [ApphudUnityWrapper paywallsDidLoadCallbackWithMaxAttempts:maxAttempts callback:^(NSString * _Nonnull paywalls, NSString * _Nullable error) {
+    [ApphudUnityAPIWrapper paywallsDidLoadCallbackWithMaxAttempts:maxAttempts callback:^(NSString * _Nonnull paywalls, NSString * _Nullable error) {
         SendCallback2ToUnity(callback, paywalls, error);
     }];
 }
@@ -200,7 +114,7 @@ extern "C" {
 #endif
 
 const char * ApphudUnity_subscriptions(void) {
-    return makeStringCopy(cstringFromString([ApphudUnityWrapper subscriptions]));
+    return makeStringCopy(cstringFromString([ApphudUnityAPIWrapper subscriptions]));
 }
 
 #ifdef __cplusplus
@@ -212,7 +126,7 @@ extern "C" {
 #endif
 
 const char * ApphudUnity_nonRenewingPurchases(void) {
-    return makeStringCopy(cstringFromString([ApphudUnityWrapper nonRenewingPurchases]));
+    return makeStringCopy(cstringFromString([ApphudUnityAPIWrapper nonRenewingPurchases]));
 }
 
 #ifdef __cplusplus
@@ -225,7 +139,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_paywallShown(const char *paywallIdentifier, const char *placementIdentifier) {
-    [ApphudUnityWrapper paywallShownWithIdentifier:cstringToString(paywallIdentifier) placementIdentifier:cstringToString(placementIdentifier)];
+    [ApphudUnityAPIWrapper paywallShownWithIdentifier:cstringToString(paywallIdentifier) placementIdentifier:cstringToString(placementIdentifier)];
 }
 
 #ifdef __cplusplus
@@ -237,7 +151,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_paywallClosed(const char *paywallIdentifier, const char *placementIdentifier) {
-    [ApphudUnityWrapper paywallClosedWithIdentifier:cstringToString(paywallIdentifier) placementIdentifier:cstringToString(placementIdentifier)];
+    [ApphudUnityAPIWrapper paywallClosedWithIdentifier:cstringToString(paywallIdentifier) placementIdentifier:cstringToString(placementIdentifier)];
 }
 
 #ifdef __cplusplus
@@ -249,7 +163,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_purchase(const char *productId,const char *placementIdentifier,const char *paywallIdentifier, UnityAction callback) {
-    [ApphudUnityWrapper purchaseWithProductId:cstringToString(productId) placementIdentifier:cstringToString(placementIdentifier) paywallIdentifier:cstringToString(paywallIdentifier) callback:^(NSString * _Nonnull purchaseResult) {
+    [ApphudUnityAPIWrapper purchaseWithProductId:cstringToString(productId) placementIdentifier:cstringToString(placementIdentifier) paywallIdentifier:cstringToString(paywallIdentifier) callback:^(NSString * _Nonnull purchaseResult) {
         SendCallbackToUnity(callback, purchaseResult);
     }];
 }
@@ -259,7 +173,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_restorePurchases(UnityAction callback) {
-    [ApphudUnityWrapper restorePurchasesWithCallback:^(NSString * _Nullable subscriptions, NSString * _Nullable nonRenewingPurchases, NSString * _Nullable error) {
+    [ApphudUnityAPIWrapper restorePurchasesWithCallback:^(NSString * _Nullable subscriptions, NSString * _Nullable nonRenewingPurchases, NSString * _Nullable error) {
         SendCallback3ToUnity(callback, subscriptions, nonRenewingPurchases, error);
     }];
 }
@@ -312,7 +226,7 @@ extern "C" {
 #endif
 
 bool ApphudUnity_isNonRenewingPurchaseActive(const char *productId) {
-    return [ApphudUnityWrapper isNonRenewingPurchaseActiveWithProductIdentifier:cstringToString(productId)];
+    return [ApphudUnityAPIWrapper isNonRenewingPurchaseActiveWithProductIdentifier:cstringToString(productId)];
 }
 
 #ifdef __cplusplus
@@ -360,7 +274,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_setUserProperty(const char *key, const char *valueJson, bool setOnce ) {
-    [ApphudUnityWrapper setUserPropertyWithKey:cstringToString(key) valueJson:cstringToString(valueJson) setOnce:setOnce];
+    [ApphudUnityAPIWrapper setUserPropertyWithKey:cstringToString(key) valueJson:cstringToString(valueJson) setOnce:setOnce];
 }
 
 #ifdef __cplusplus
@@ -373,7 +287,7 @@ extern "C" {
 #endif
 
 void ApphudUnity_incrementUserProperty(const char *key, const char *byJson) {
-    [ApphudUnityWrapper incrementUserPropertyWithKey:cstringToString(key) byJson:cstringToString(byJson)];
+    [ApphudUnityAPIWrapper incrementUserPropertyWithKey:cstringToString(key) byJson:cstringToString(byJson)];
 }
 
 #ifdef __cplusplus
@@ -384,8 +298,24 @@ void ApphudUnity_incrementUserProperty(const char *key, const char *byJson) {
 extern "C" {
 #endif
 
-void ApphudUnity_addAttribution(const char *provider, const char *dataJson, const char *identifer) {
-    [ApphudUnityWrapper addAttributionWithProvider:cstringToString(provider) dataJson:cstringToString(dataJson) identifer:cstringToString(identifer)];
+void ApphudUnity_addAttribution(const char *provider, const char *dataJson, const char *identifer, UnityAction callback) {
+    [ApphudUnityAPIWrapper addAttributionWithProvider:cstringToString(provider) dataJson:cstringToString(dataJson) identifer:cstringToString(identifer) callback:^(BOOL value) {
+        SendCallbackBoolToUnity(callback, value);
+    }];
+}
+
+#ifdef __cplusplus
+}
+#endif
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void ApphudUnity_loadFallbackPaywallsWithCallback(UnityAction callback) {
+    [ApphudUnityAPIWrapper loadFallbackPaywallsWithCallback: ^(NSString * _Nonnull paywalls, NSString * _Nullable error) {
+        SendCallback2ToUnity(callback, paywalls, error);
+    }];
 }
 
 #ifdef __cplusplus
@@ -409,7 +339,45 @@ extern "C" {
 #endif
 
 void ApphudUnity_setHeaders() {
-    [ApphudUnityWrapper setHeaders];
+    [ApphudUnityAPIWrapper setHeaders];
+}
+
+#ifdef __cplusplus
+}
+#endif
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void ApphudUnity_setPaywallsCacheTimeout(double value) {
+    [Apphud setPaywallsCacheTimeout:value];
+}
+
+#ifdef __cplusplus
+}
+#endif
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void ApphudUnity_submitPushNotificationsTokenString(const char *string, UnityAction callback) {
+    [Apphud submitPushNotificationsTokenStringWithString:cstringToString(string) callback:^(BOOL value) {
+        SendCallbackBoolToUnity(callback, value);
+    }];
+}
+
+#ifdef __cplusplus
+}
+#endif
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void ApphudUnity_willPurchaseProductFrom(const char *paywallIdentifier, const char *placementIdentifier) {
+    [Apphud willPurchaseProductFromPaywallIdentifier:cstringToString(paywallIdentifier) placementIdentifier:cstringToString(placementIdentifier)];
 }
 
 #ifdef __cplusplus