+++ Runtime/SDK/Common/IApphudSDK.cs
@@ -8,8 +8,8 @@ namespace Apphud.Unity.Common
     {
         string UserId { get; }
         string DeviceId { get; }
-        void Start(string apiKey, Action<ApphudUser> callback);
-        void Start(string apiKey, string userId, Action<ApphudUser> callback);
+        void Start(string apiKey, Action<ApphudUser> callback, bool observerMode);
+        void Start(string apiKey, string userId, Action<ApphudUser> callback, bool observerMode);
         void LogOut();
         void UpdateUserId(string userId);
 
@@ -32,12 +32,22 @@ namespace Apphud.Unity.Common
         void SetUserProperty(ApphudUserPropertyKey key, object value, bool setOnce);
         void IncrementUserProperty(ApphudUserPropertyKey key, object by);
         void AddAttribution(ApphudAttributionProvider provider, Dictionary<string, object> data = null, string identifier = null);
+#if APPHUD_FB
+        void AddFacebookAttribution(Action<string> onError = null);
+#endif
+        void LoadFallbackPaywalls(Action<List<ApphudPaywall>, ApphudError> callback);
+        void InvalidatePaywallsCache();
 
 #if UNITY_ANDROID
+        bool IsFallbackMode();
         void RefreshUserData();
         void CollectDeviceIdentifiers();
+        void TrackPurchase(string productID, string offerIdToken, string paywallIdentifier, string placementIdentifier);
 #elif UNITY_IOS
+        void TrackAppleSearchAds();
+        void WillPurchaseProductFrom(string paywallIdentifier, string placementIdentifier);
         void SetDeviceIdentifiers(string idfa, string idfv);
+        void SubmitPushNotificationsTokenString(string str, Action<bool> callback);
 #endif
     }
 }