+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [1.1.0] - 2024-07-29
+
+- Observer mode
+- Editor mode exceptions handling
+- New methods
+  - LoadFallbackPaywalls
+  - AddFacebookAttribution
+  - InvalidatePaywallsCache
+- New Android methods
+  - IsFallbackMode
+  - TrackPurchase
+- New iOS methods
+  - TrackAppleSearchAds
+  - WillPurchaseProductFrom
+  - SubmitPushNotificationsTokenString
+
 ## [1.0.0] - 2024-05-16
 
 ### This is the first release of _Apphud Unity SDK_.